// src/app/models/dtos/e-dtos/sal-order-detail.dto.ts
import { SALOrderDetailDeliveryStatusEnum } from "../../enums/e-status/sal-order-detail-delivery-status.enum";
import { SALOrderDetailStatusEnum } from "../../enums/e-status/sal-order-detail-status.enum";
import { SALOrderMasterStatusRetailEnum } from "../../enums/e-status/sal-order-master-status-retail.enum";
import { SALOrderDetailPaymentTypeEnum } from "../../enums/e-type/sal-order-detail-payment-type.enum";
import { SALOrderDetailTypeDataEnum } from "../../enums/e-type/sal-order-detail-type-data.enum";

class SALOrderDetailDTO {
  Code: number = 0;
  Master: number;
  CSVehicle: number;
  Price: number = 0;
  RegisterDate: Date | null;
  RegisterStaff: number;
  PlateDeliveryDate: Date | null;
  TechnicalStaff: number;
  DeliveryStatus: SALOrderDetailDeliveryStatusEnum = SALOrderDetailDeliveryStatusEnum.NEW;
  CreateBy: string;
  CreateTime?: Date | null;
  LastModifiedBy: string;
  LastModifiedTime?: Date | null;
  TypeData: number = SALOrderDetailTypeDataEnum.BUY;
  OrderBy: number;
  VehicleColor: number;
  Status: SALOrderDetailStatusEnum;
  IsPaid: boolean;
  Owner: number;
  PaymentType: SALOrderDetailPaymentTypeEnum;
  FinanceCompany: number;
  InstallmentTerm: number;
  InterestRate: number;
  ContractID: string = '';
  TypeOfVehicleName: string = '';
  VehicleName: string = '';
  VehicleColorName: string = '';
  IsCompare: boolean = false;
  IsFollow: boolean = false;
  IsLock: boolean = false;
  Warehouse: number;
  HeadTransfer: number;
  DepositAmount: number;
}

export class SALOrderDetailCusDTO extends SALOrderDetailDTO {
  ID: string = '';
  MasterStatus: SALOrderMasterStatusRetailEnum;
  TypeOfVehicle: number;
  Vehicle: number;
  SaleDate: Date | null;
  FrameSeri: string = '';
  EngineSeri: string = '';
  PlateNo: string = '';
  CustomerName: string = '';
  Phone: string = '';
  Address: string = '';
  AmountPaid: number = 0;
  SaleStaffName: string = '';
  RegisterStaffName: string = '';
  TechnicalStaffName: string = '';
  MasterStatusName: string = '';
  PaymentMethodName: string = '';
  PaymentMethod: number;
  VehicleImage: string = '';
  InsuranceNumber: string = '';
  InsurancePeriod: number = 0;
  InsuranceTime: number = 0;
  Version: string = '';
  Cylinder: string = '';
  DeliveryStatusName: string = '';
  Rescue: number;
  Zalo: string = '';
  MyHonda: number;
  MyHondaReason: number;
  MyHondaReasonDetail: string = '';
  OA: string = '';
  OAReason: number;
  OAReasonDetail: number;
  Personality: string = '';
  Interest: string = '';
  OtherConsiderations: string = '';
  TypeOfPart: number;
  TypeOfPartSpecs: number;
  Quantity: number = 0;
  UnitPrice: number = 0;
  Remark: string = '';
  ImageSetting1: string = '';
  TotalService: number = 0;
  TotalPart: number = 0;
  TotalPromotion: number = 0;
  StatusName: string = '';
  Customer: number;
  TransferFrameSeri: string = '';
  TransferEngineSeri: string = '';
  OrderQuantity: number = 1;
  OrderTypeData: number;
  IsOrderLock: boolean = false;
  IsChecked: boolean = false;
  ListStock = [];
  SALOrderDetailPromotion: number;
  WHOut: number;
  WHIn: number;
  TransferCSVehicle: number;
}